Fix Segmented Display crashing server
